Move processingPaths to src/config/ to respect pipeline/game boundary (#115)
processingPaths() is a pure path builder with no pipeline logic. Game
code (resolve-session, search-content, game-engine) was importing it
from src/content/, violating the "content pipeline is separate from
game code" rule in CLAUDE.md.

- Canonical location: src/config/processing-paths.ts
- src/content/processing-paths.ts becomes a re-export so internal
  pipeline imports continue working unchanged
- Game code imports updated to src/config/

Closes #115
