+++ src/content/processing-paths.ts
@@ -1,50 +1,5 @@
 /**
- * Path builders for the content processing pipeline output.
- *
- * All processed output goes to ~/.machine-violet/systems/<collection-slug>/.
- * Parallels ingestPaths() from job-manager.ts.
+ * Re-export from canonical location in src/config/.
+ * Content pipeline code can continue importing from here.
  */
-
-import { join } from "node:path";
-
-/**
- * Build paths for processed output of a collection.
- *
- * @param homeDir - Application home directory (e.g. ~/.machine-violet).
- * @param collectionSlug - Slugified collection name (e.g. "d-d-5e").
- */
-export function processingPaths(homeDir: string, collectionSlug: string) {
-  const base = join(homeDir, "systems", collectionSlug);
-  return {
-    /** Root of processed output for this collection. */
-    base,
-    /** Pipeline state file. */
-    pipelineState: join(base, "pipeline.json"),
-    /** Stage 1 output — content catalog. */
-    catalog: join(base, "catalog.json"),
-    /** Stage 2 output — draft entities root. */
-    draftsDir: join(base, "drafts"),
-    /** Stage 2 output — draft entity file. */
-    draftFile: (category: string, slug: string) =>
-      join(base, "drafts", category, `${slug}.md`),
-    /** Stage 2 output — draft category directory. */
-    draftCategoryDir: (category: string) =>
-      join(base, "drafts", category),
-    /** Stage 3 output — merged entities root. */
-    entitiesDir: join(base, "entities"),
-    /** Stage 3 output — merged entity file. */
-    entityFile: (category: string, slug: string) =>
-      join(base, "entities", category, `${slug}.md`),
-    /** Stage 3 output — entity category directory. */
-    entityCategoryDir: (category: string) =>
-      join(base, "entities", category),
-    /** Stage 4 output — index. */
-    index: join(base, "index.md"),
-    /** Stage 4 output — cheat sheet. */
-    cheatSheet: join(base, "cheat-sheet.md"),
-    /** Stage 4 output — faceted entity index per category. */
-    facets: (category: string) => join(base, "entities", category, "facets.json"),
-    /** Stage 5 output — rule card. */
-    ruleCard: join(base, "rule-card.md"),
-  };
-}
+export { processingPaths } from "../config/processing-paths.js";
